+++ src/components/AuthoritiesLookup/AuthoritiesLookup.js
@@ -7,8 +7,12 @@ import { useIntl } from 'react-intl';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 
-import { Pane } from '@folio/stripes/components';
+import {
+  Pane,
+  PaneMenu,
+} from '@folio/stripes/components';
 import { AppIcon } from '@folio/stripes/core';
+import { ExpandFilterPaneButton } from '@folio/stripes/smart-components';
 import {
   AuthoritiesSearchContext,
   AuthoritiesSearchPane,
@@ -18,6 +22,7 @@ import {
   SearchResultsList,
   SelectedAuthorityRecordContext,
   navigationSegments,
+  useAutoOpenDetailView,
 } from '@folio/stripes-authority-components';
 
 import MarcAuthorityView from '../MarcAuthorityView';
@@ -64,6 +69,7 @@ const AuthoritiesLookup = ({
   const intl = useIntl();
   const [isFilterPaneVisible, setIsFilterPaneVisible] = useState(true);
   const [showDetailView, setShowDetailView] = useState(false);
+  const [itemToView, setItemToView] = useState(null);
 
   const {
     filters,
@@ -88,6 +94,13 @@ const AuthoritiesLookup = ({
     setShowDetailView(true);
   };
 
+  const handleRowFocus = ({ selector, localClientTop }) => {
+    setItemToView({
+      selector,
+      localClientTop,
+    });
+  };
+
   useEffect(() => {
     closeDetailView();
     // eslint-disable-next-line
@@ -103,6 +116,8 @@ const AuthoritiesLookup = ({
     onSubmitSearch(e, ...rest);
   };
 
+  useAutoOpenDetailView(authorities, openDetailView);
+
   const renderPaneSub = () => {
     if (navigationSegmentValue === navigationSegments.browse) {
       return null;
@@ -114,6 +129,20 @@ const AuthoritiesLookup = ({
           totalRecords,
         })}
       </span>
+    );
+  };
+
+  const renderResultsFirstMenu = () => {
+    if (isFilterPaneVisible) {
+      return null;
+    }
+
+    return (
+      <PaneMenu>
+        <ExpandFilterPaneButton
+          onClick={toggleFilterPane}
+        />
+      </PaneMenu>
     );
   };
 
@@ -131,12 +160,12 @@ const AuthoritiesLookup = ({
   const renderResultList = () => (
     <Pane
       id="authority-search-results-pane"
-      className={classNames(css.pane, css.focusIndicator, { [css.hidden]: showDetailView })}
       data-testid="authority-search-results-pane"
       appIcon={<AppIcon app="marc-authorities" />}
       defaultWidth="fill"
       paneTitle={intl.formatMessage({ id: 'stripes-authority-components.meta.title' })}
       paneSub={renderPaneSub()}
+      firstMenu={renderResultsFirstMenu()}
       padContent={false}
       noOverflow
       height={MAIN_PANE_HEIGHT}
@@ -159,7 +188,8 @@ const AuthoritiesLookup = ({
         hasPrevPage={hasPrevPage}
         hidePageIndices={hidePageIndices}
         renderHeadingRef={renderHeadingRef}
-        onOpenRecord={openDetailView}
+        itemToView={itemToView}
+        onRowFocus={handleRowFocus}
       />
     </Pane>
   );
@@ -177,12 +207,14 @@ const AuthoritiesLookup = ({
         excludedBrowseFilters={excludedFilters}
         onShowDetailView={setShowDetailView}
       />
-      {showDetailView &&
-        <MarcAuthorityView
-          onCloseDetailView={closeDetailView}
-        />
+      {showDetailView
+        ? (
+          <MarcAuthorityView
+            onCloseDetailView={closeDetailView}
+          />
+        )
+        : renderResultList()
       }
-      {renderResultList()}
     </>
   );
 };